requirements.txt
fastapi~=0.115.8
pydantic~=2.10.6
python-dotenv~=1.0.1
langchain~=0.3.17
requests~=2.32.3
langchain-community~=0.3.16
langchain-openai~=0.3.3